FEM: beam sectiong and element geometry 1D object, make name fit in FEM name system
