+++ src/Mod/Fem/App/CMakeLists.txt
@@ -89,7 +89,7 @@ SET(FemScripts_SRCS
 
 SET(FemObjectsScripts_SRCS
     PyObjects/__init__.py
-    PyObjects/_FemBeamSection.py
+    PyObjects/_FemElementGeometry1D.py
     PyObjects/_FemConstraintSelfWeight.py
     PyObjects/_FemFluidSection.py
     PyObjects/_FemMaterialMechanicalNonlinear.py
@@ -107,7 +107,7 @@ SET(FemGuiScripts_SRCS
     PyGui/FemCommands.py
     PyGui/__init__.py
     PyGui/_CommandFemAnalysis.py
-    PyGui/_CommandFemBeamSection.py
+    PyGui/_CommandFemElementGeometry1D.py
     PyGui/_CommandFemConstraintSelfWeight.py
     PyGui/_CommandFemFluidSection.py
     PyGui/_CommandFemMaterialMechanicalNonlinear.py
@@ -127,7 +127,7 @@ SET(FemGuiScripts_SRCS
     PyGui/_CommandFemSolverControl.py
     PyGui/_CommandFemSolverRun.py
     PyGui/_CommandFemSolverZ88.py
-    PyGui/_TaskPanelFemBeamSection.py
+    PyGui/_TaskPanelFemElementGeometry1D.py
     PyGui/_TaskPanelFemFluidSection.py
     PyGui/_TaskPanelFemMaterial.py
     PyGui/_TaskPanelFemMeshGmsh.py
@@ -136,7 +136,7 @@ SET(FemGuiScripts_SRCS
     PyGui/_TaskPanelFemResultShow.py
     PyGui/_TaskPanelFemShellThickness.py
     PyGui/_TaskPanelFemSolverCalculix.py
-    PyGui/_ViewProviderFemBeamSection.py
+    PyGui/_ViewProviderFemElementGeometry1D.py
     PyGui/_ViewProviderFemConstraintSelfWeight.py
     PyGui/_ViewProviderFemFluidSection.py
     PyGui/_ViewProviderFemMaterial.py
@@ -148,7 +148,7 @@ SET(FemGuiScripts_SRCS
     PyGui/_ViewProviderFemShellThickness.py
     PyGui/_ViewProviderFemSolverCalculix.py
     PyGui/_ViewProviderFemSolverZ88.py
-    PyGui/TaskPanelFemBeamSection.ui
+    PyGui/TaskPanelFemElementGeometry1D.ui
     PyGui/TaskPanelFemFluidSection.ui
     PyGui/TaskPanelFemMaterial.ui
     PyGui/TaskPanelFemMeshGmsh.ui
