+++ src/Mod/Fem/FemMeshTools.py
@@ -404,7 +404,7 @@ def get_femelement_sets_from_group_data(femmesh, fem_objects):
 def get_elset_short_name(obj, i):
     if hasattr(obj, "Proxy") and obj.Proxy.Type == 'FemMaterial':
         return 'Mat' + str(i)
-    elif hasattr(obj, "Proxy") and obj.Proxy.Type == 'FemBeamSection':
+    elif hasattr(obj, "Proxy") and obj.Proxy.Type == 'FemElementGeometry1D':
         return 'Beam' + str(i)
     elif hasattr(obj, "Proxy") and obj.Proxy.Type == 'FemFluidSection':
         return 'Fluid' + str(i)
@@ -1030,7 +1030,7 @@ def get_analysis_group_elements(aAnalysis, aPart):
         else:
             FreeCAD.Console.PrintError('Problem: more than one object with empty references.\n')
             print('We gone try to get the empty material references anyway.\n')
-            # ShellThickness, BeamSection and FluidSection could have empty references, but on solid meshes only materials should have empty references
+            # ShellThickness, ElementGeometry1D and FluidSection could have empty references, but on solid meshes only materials should have empty references
             for er in empty_references:
                 print(er.Name)
             group_elements = get_anlysis_empty_references_group_elements(group_elements, aAnalysis, aPart.Shape)
