+++ src/Mod/Fem/Gui/Workbench.cpp
@@ -77,7 +77,7 @@ Gui::ToolBarItem* Workbench::setupToolBars() const
          << "FEM_MaterialSolid"
          << "FEM_MaterialFluid"
          << "FEM_MaterialMechanicalNonlinear"
-         << "FEM_BeamSection"
+         << "FEM_ElementGeometry1D"
          << "FEM_ShellThickness"
          << "FEM_FluidSection"
          << "Separator"
@@ -152,7 +152,7 @@ Gui::MenuItem* Workbench::setupMenuBar() const
          << "FEM_MaterialSolid"
          << "FEM_MaterialFluid"
          << "FEM_MaterialMechanicalNonlinear"
-         << "FEM_BeamSection"
+         << "FEM_ElementGeometry1D"
          << "FEM_ShellThickness"
          << "FEM_FluidSection"
          << "Separator"
