+++ src/Mod/Fem/PyGui/TaskPanelFemElementGeometry1D.ui
